test: Unit tests for nitro enclave NSM device behaviour
This is useful for testing QEMU's nitro-enclave machine's NSM emulation.
The tests have been written by testing out real NSM in AWS nitro enclave
VM.

Build a docker image using the Dockerfile and then use nitro-cli to build
an EIF file. Take note of the measurements from the nitro-cli stdout.
Run the EIF file in AWS nitro enclave VM in debug mode and attach a console
to see the VM's output that the tests pass. Then run the EIF file using
QEMU and verify from the output that the measurements match and that the
tests pass too.